+++ example/calculator.dart
@@ -713,6 +713,4 @@ class State<T> {
 
     return result;
   }
-
-  static T as<T>(T? value) => value as T;
 }