+++ lib/src/capture/captured_value_action.dart
@@ -1,5 +1,6 @@
 part of '../../capture.dart';
 
+@experimental
 class CapturedValueAction extends SemanticAction<int> {
   final Object key;
 
@@ -7,7 +8,7 @@ class CapturedValueAction extends SemanticAction<int> {
 
   @override
   String build(Context context, String name, List<String> arguments) {
-    final value = context.getSemanticValue(key);
+    final value = context.getCapturedValue(key);
     return replaceParameters(value.value, const [], arguments);
   }
 }