+++ lib/src/parser_builder/named.dart
@@ -53,12 +53,12 @@ class Named<I, O> extends ParserBuilder<I, O> {
     final localDeclarations = context.localDeclarations;
     final localRegistry = context.localRegistry;
     final refersToStateSource = context.refersToStateSource;
-    final semanticValues = context.semanticValues;
+    final semanticValues = context.capturedValues;
     context.localAllocator = localAllocator.clone();
     context.localDeclarations = {};
     context.localRegistry = {};
     context.refersToStateSource = false;
-    context.semanticValues = {};
+    context.capturedValues = {};
     if (parser.getResultType() == 'void') {
       fast = true;
     }
@@ -98,7 +98,7 @@ class Named<I, O> extends ParserBuilder<I, O> {
     }
 
     buffer.writeln('}');
-    context.semanticValues = semanticValues;
+    context.capturedValues = semanticValues;
     context.refersToStateSource = refersToStateSource;
     context.localDeclarations = localDeclarations;
     context.localRegistry = localRegistry;