+++ lib/src/parser_builder/runtime.dart
@@ -419,8 +419,6 @@ class State<T> {
 
     return result;
   }
-
-  static T as<T>(T? value) => value as T;
 }''';
 
   static const _classStateNoMemo = r'''
@@ -625,8 +623,6 @@ class State<T> {
 
     return result;
   }
-
-  static T as<T>(T? value) => value as T;
 }''';
 
   static const _extensionString = r'''