+++ lib/src/parser_builder/semantic_value.dart
@@ -12,6 +12,6 @@ class SemanticValue<T> {
       return name;
     }
 
-    return 'State.as<$T>($name)';
+    return '($name as dynamic) as $T';
   }
 }