+++ test/_json_number_parser.dart
@@ -671,6 +671,4 @@ class State<T> {
 
     return result;
   }
-
-  static T as<T>(T? value) => value as T;
 }